requirements/common.txt
django>=1.5
django-cms>=2.4
djangocms-text-ckeditor>=2.0
Pillow
pytz
